+++ test/TestHelpers/SimStateContainer.h
@@ -17,11 +17,11 @@ namespace TestHelpers
 struct SimStateContainerData
 {
   int state_id;  // The simulation state to which the cell belongs
-  long long cell_id;  // The unique cell identifier
+  unsigned long long cell_id;  // The unique cell identifier
   double cell_energy;  // The deposited energy in that cell
   long long layer;  // The layer ID to which the cell belons
-  double eta, phi, r, x, y, z;  // The center positions of the cell
-  double deta, dphi, dr, dx, dy, dz;  // The dimensions of the cell
+  float eta, phi, r, x, y, z;  // The center positions of the cell
+  float deta, dphi, dr, dx, dy, dz;  // The dimensions of the cell
   bool isBarrel;  // Whether the cell belongs to a barrel or endcap layer
   bool isXYZ, isEtaPhiR, isEtaPhiZ, isRPhiZ;  // The coordinate system in which
                                               // cell dimensions are described
@@ -61,10 +61,10 @@ class SimStateContainer
     for (const auto& state : m_container) {
       TFCSSimulationState::cellmap cellmap = state.cells();
       // First order the cellmap by cell id
-      std::map<long long, float> ordered_cellmap(cellmap.begin(),
-                                                 cellmap.end());
+      std::map<unsigned long long, float> ordered_cellmap(cellmap.begin(),
+                                                          cellmap.end());
       for (const auto& cell_map_iter : ordered_cellmap) {
-        long long cell_id = cell_map_iter.first;
+        unsigned long long cell_id = cell_map_iter.first;
         const float cell_energy = cell_map_iter.second;
 
         // Get the cell
